Atualização do projeto para o funcionamento proposto.
- Adequação do nome do package para lowercase
