--- src/main/java/com/juan/orangeTalents/controleUsuarios/exception/EmailExixtenteException.java
@@ -1,5 +1,0 @@
-package com.juan.orangeTalents.controleUsuarios.exception;
-
-public class EmailExixtenteException {
-
-}
