+++ src/main/java/com/juan/orangeTalents/controleusuarios/exception/CpfExistenteException.java
@@ -1,0 +1,5 @@
+package com.juan.orangeTalents.controleusuarios.exception;
+
+public class CpfExistenteException {
+
+}
