+++ src/main/java/com/juan/orangeTalents/controleusuarios/exception/UsuarioNaoEncontradoException.java
@@ -1,0 +1,5 @@
+package com.juan.orangeTalents.controleusuarios.exception;
+
+public class UsuarioNaoEncontradoException {
+
+}
