+++ src/main/java/com/juan/orangeTalents/controleusuarios/service/EnderecoService.java
@@ -1,10 +1,10 @@
-package com.juan.orangeTalents.controleUsuarios.service;
+package com.juan.orangeTalents.controleusuarios.service;
 
 import java.util.List;
 import java.util.Optional;
 
-import com.juan.orangeTalents.controleUsuarios.dto.EnderecoDto;
-import com.juan.orangeTalents.controleUsuarios.model.Endereco;
+import com.juan.orangeTalents.controleusuarios.dto.EnderecoDto;
+import com.juan.orangeTalents.controleusuarios.model.Endereco;
 
 //@FeignClient(url = "viacep.com.br/ws/", name = "viacepService")
 public interface EnderecoService {
@@ -14,12 +14,12 @@ public interface EnderecoService {
 	List<Endereco> getAllEnderecos();
 
 	Endereco insertEndereco(EnderecoDto e);
-	
+
 //	List<Endereco> getUsuariosDeEnderecos();
-	
+
 //	@GetMapping("{cep}/json")
 //    Endereco buscaEnderecoPorCep(@PathVariable("cep") String cep);
-//	
+//
 //	@RequestMapping(value = "/rest/insertEndereco", method = RequestMethod.POST)
 //	void insertEndereco(Endereco endereco);
 }
