+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2
 
 #Datasource
-spring.datasource.url=jdbc:h2:/home/juan/Documentos/db_spring
+spring.datasource.url=jdbc:h2:~/Documents/
 spring.datasource.username=spring
 spring.datasource.password=spring
 spring.datasource.driver-class-name=org.h2.Driver
